MainActivity is now an entry point to test the different WM artifacts. CodelabActivity is now where the codelab sample lives.
