+++ window-manager/src/main/res/layout/activity_main.xml
@@ -1,72 +1,35 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ /*
-  ~  * Copyright (C) 2021 Google Inc.
-  ~  *
-  ~  * Licensed under the Apache License, Version 2.0 (the "License");
-  ~  * you may not use this file except in compliance with the License.
-  ~  * You may obtain a copy of the License at
-  ~  *
-  ~  *     http://www.apache.org/licenses/LICENSE-2.0
-  ~  *
-  ~  * Unless required by applicable law or agreed to in writing, software
-  ~  * distributed under the License is distributed on an "AS IS" BASIS,
-  ~  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  * See the License for the specific language governing permissions and
-  ~  * limitations under the License.
-  ~  */
-  -->
-
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/constraint_layout"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    android:gravity="center"
+    android:orientation="vertical">
 
-    <TextView
-        android:id="@+id/window_metrics"
+    <Button
+        android:id="@+id/codelab_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:padding="20dp"
-        android:text="@string/window_metrics"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toTopOf="@+id/layout_change"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed" />
+        android:layout_gravity="start"
+        android:layout_margin="20dp"
+        android:text="Codelab sample"
+        android:textSize="30sp" />
 
-    <TextView
-        android:id="@+id/layout_change"
+    <Button
+        android:id="@+id/flow_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:padding="20dp"
-        android:text="@string/layout_change_text"
-        android:textSize="20sp"
-        app:layout_constrainedWidth="true"
-        app:layout_constraintBottom_toTopOf="@+id/configuration_changed"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/window_metrics" />
+        android:layout_gravity="start"
+        android:layout_margin="20dp"
+        android:text="Flow sample"
+        android:textSize="30sp" />
 
-    <TextView
-        android:id="@+id/configuration_changed"
+    <Button
+        android:id="@+id/rxjava_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:padding="20dp"
-        android:text="@string/configuration_changed"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/layout_change" />
+        android:layout_gravity="start"
+        android:layout_margin="20dp"
+        android:text="RxJava sample"
+        android:textSize="30sp" />
 
-    <!-- It's not important where this view is placed by default, it will be adjusted dynamically in runtime -->
-    <View
-        android:id="@+id/device_feature"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:background="@android:color/holo_red_dark"
-        android:visibility="gone"
-        tools:ignore="MissingConstraints" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
